% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{The data to be analyzed}
}
\value{
This function returns a dataframe.
}
\description{
This is a simple function that cleans data
}
\examples{
library(stringr)
library(dplyr)
library(lubridate)

datapath <- system.file("extdata", "earthquakes.txt", package = "gsedr")
after2000data <- readr::read_delim(datapath, delim = "\\t") \%>\% eq_clean_data()  \%>\% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000)


}
