% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
This function returns a dataframe.
}
\description{
This is a simple function that cleans data
}
\examples{
library(stringr)
library(dplyr)
library(lubridate)
datapath <- system.file("extdata", "earthquakes.txt", package = "gsedr")
after2000data <- readr::read_delim(datapath, delim = "\\t") \%>\% eq_clean_data()  \%>\% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000)


}
