% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{eq_map}
\alias{eq_map}
\title{This is a simple funtion for mapping}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Data}

\item{annot_col}{A character, which is a column in data to be shown in map}
}
\value{
Leaflet map
}
\description{
This is a simple funtion for mapping
}
\examples{
library(ggplot2)
library(leaflet)

datapath <- system.file("extdata", "earthquakes.txt", package = "gsedr")
data <- readr::read_delim(datapath, delim = "\\t") \%>\% eq_clean_data()
dddd <- data \%>\% dplyr::filter(COUNTRY \%in\% c('INDONESIA')) \%>\% dplyr::filter(!is.na(EQ_PRIMARY), !is.na(DEATHS))
dddd \%>\% dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col = "popup_text")

}
